frontend/requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-WTF==1.2.1
email-validator==2.1.0.post1
python-dotenv==1.0.1
Pillow==10.3.0
SQLAlchemy==2.0.23
Werkzeug==3.1.3